quant-fund-brazil/requirements.txt
# Dependências do Projeto de Fundo Quantitativo

# Core
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Dados de Mercado
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Backtesting
backtrader>=1.9.78
vectorbt>=0.26.0

# Análise Estatística
statsmodels>=0.14.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.28.0

# Gestão de Risco
quantlib>=1.31

# Utilitários
tqdm>=4.65.0
requests>=2.31.0
